i97: Move doppler config under firebase dir
