+++ firebase/functions/index.js
@@ -1,7 +1,7 @@
 const admin = require('firebase-admin');
 const functions = require('firebase-functions');
 const serviceAccount = require('./config/serviceAccount.json');
-const config = require('./config/env.config');
+const config = require('./config');
 const storeOauthTokens = require('./auth/storeOauthTokens');
 const createShout = require('./shouts/createShout');
 const checkTwitterGeo = require('./shouts/checkTwitterGeo');
